Remove custom archive cache file entirely
